# contract/openapi.yaml
openapi: 3.0.3
info:
  title: Booking Calls API
  description: >
    API для приложения предварительной записи на телефонный разговор.
    Владелец публикует доступные интервалы (рабочие часы на конкретные дни),
    система нарезает их на слоты по 30 минут. Гости выбирают свободный слот
    в пределах скользящего окна записи (14 дней от текущей даты) и бронируют его.
    Авторизации нет. Хранение данных — в памяти backend-процесса.
  version: "1.0.0"
servers:
  - url: /api
    description: Base path, проксируется на backend через nginx
...
paths:
...
  /owner/availability:
    get:
      operationId: listOwnerAvailability
      tags: [owner]
      summary: Список опубликованных владельцем интервалов доступности
      description: >
        Без авторизации, согласно ТЗ. Используется экраном владельца для
        просмотра уже опубликованных интервалов.
      responses:
        "200":
          description: Список интервалов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilityInterval"
    post:
      operationId: createOwnerAvailability
      tags: [owner]
      summary: Опубликовать доступный интервал (рабочие часы на день)
      description: >
        Из интервала система нарезает слоты по 30 минут. Интервал должен
        укладываться в окно записи `[now, now + 14 дней]`, иначе — `400`.
        Пересекающиеся с уже опубликованными интервалы — `409`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilityIntervalRequest"
      responses:
        "201":
          description: Интервал опубликован, слоты нарезаны
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityInterval"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Интервал пересекается с уже опубликованным
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  responses:
    BadRequest:
      description: Некорректные параметры запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    AvailabilityInterval:
      type: object
      required: [id, date, startTime, endTime]
      properties:
        id:
          type: string
          example: "avail_20260901_0900_1300"
        date:
          type: string
          format: date
          description: Дата, на которую опубликован интервал (UTC)
        startTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          description: Время начала интервала в формате HH:MM (UTC)
          example: "09:00"
        endTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          description: Время конца интервала в формате HH:MM (UTC)
          example: "13:00"
    CreateAvailabilityIntervalRequest:
      type: object
      required: [date, startTime, endTime]
      properties:
        date:
          type: string
          format: date
        startTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "09:00"
        endTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "13:00"
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Машиночитаемый код ошибки
          example: "slot_already_booked"
        message:
          type: string
          example: "Слот уже забронирован"